requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# HTTP & API
requests==2.31.0
httpx==0.25.2
tenacity==8.2.3

# Media Processing
yt-dlp==2023.12.30
opencv-python==4.8.1.78
imagehash==4.3.1
Pillow==10.1.0
librosa==0.10.1
numpy==1.24.4
scipy==1.11.4
audioread==3.0.1
pyacoustid==1.3.0

# Task Queue
celery==5.3.4
redis==5.0.1

# Storage
boto3==1.34.0
minio==7.2.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
click==8.1.7

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Linting & Formatting
ruff==0.1.6
black==23.11.0
mypy==1.7.1
pre-commit==3.6.0

# Monitoring & Metrics
prometheus-client==0.19.0
structlog==23.2.0

# Security
cryptography>=42.0.0